+++ http/server.go
@@ -51,10 +51,10 @@ import (
 	"github.com/datarhei/core/v16/session"
 	"github.com/datarhei/core/v16/srt"
 
-	mwbodysize "github.com/datarhei/core/v16/http/middleware/bodysize"
 	mwcache "github.com/datarhei/core/v16/http/middleware/cache"
 	mwcors "github.com/datarhei/core/v16/http/middleware/cors"
 	mwgzip "github.com/datarhei/core/v16/http/middleware/gzip"
+	mwhlsrewrite "github.com/datarhei/core/v16/http/middleware/hlsrewrite"
 	mwiplimit "github.com/datarhei/core/v16/http/middleware/iplimit"
 	mwlog "github.com/datarhei/core/v16/http/middleware/log"
 	mwmime "github.com/datarhei/core/v16/http/middleware/mime"
@@ -132,6 +132,7 @@ type server struct {
 		cors       echo.MiddlewareFunc
 		cache      echo.MiddlewareFunc
 		session    echo.MiddlewareFunc
+		hlsrewrite echo.MiddlewareFunc
 	}
 
 	gzip struct {
@@ -192,6 +193,12 @@ func NewServer(config Config) (Server, error) {
 		}
 
 		s.filesystems[filesystem.Name] = filesystem
+
+		if fs.Filesystem.Type() == "disk" {
+			s.middleware.hlsrewrite = mwhlsrewrite.NewHLSRewriteWithConfig(mwhlsrewrite.HLSRewriteConfig{
+				PathPrefix: fs.Filesystem.Base(),
+			})
+		}
 	}
 
 	if _, ok := corsPrefixes["/"]; !ok {
@@ -330,7 +337,6 @@ func NewServer(config Config) (Server, error) {
 			return nil
 		},
 	}))
-	s.router.Use(mwbodysize.New())
 	s.router.Use(mwsession.NewHTTPWithConfig(mwsession.HTTPConfig{
 		Collector: config.Sessions.Collector("http"),
 	}))
@@ -394,9 +400,9 @@ func (s *server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 func (s *server) setRoutes() {
 	gzipMiddleware := mwgzip.NewWithConfig(mwgzip.Config{
-		Level:        mwgzip.BestSpeed,
-		MinLength:    1000,
-		ContentTypes: []string{""},
+		Level:     mwgzip.BestSpeed,
+		MinLength: 1000,
+		Skipper:   mwgzip.ContentTypeSkipper(nil),
 	})
 
 	// API router grouo
@@ -440,9 +446,9 @@ func (s *server) setRoutes() {
 
 		if filesystem.Gzip {
 			fs.Use(mwgzip.NewWithConfig(mwgzip.Config{
-				Level:        mwgzip.BestSpeed,
-				MinLength:    1000,
-				ContentTypes: s.gzip.mimetypes,
+				Skipper:   mwgzip.ContentTypeSkipper(s.gzip.mimetypes),
+				Level:     mwgzip.BestSpeed,
+				MinLength: 1000,
 			}))
 		}
 
@@ -629,6 +635,7 @@ func (s *server) setRoutesV3(v3 *echo.Group) {
 	// v3 Log
 	v3.GET("/log", s.v3handler.log.Log)
 
-	// v3 Resources
+	// v3 Metrics
+	v3.GET("/metrics", s.v3handler.resources.Describe)
 	v3.POST("/metrics", s.v3handler.resources.Metrics)
 }