+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Core
 
+### Core v16.10.0 > v16.10.1
+
+-   Add email address in TLS config for Let's Encrypt
+-   Fix use of Let's Encrypt production CA
+
+### Core v16.9.1 > v16.10.0
+
+-   Add HLS session middleware to diskfs
+-   Add /v3/metrics (get) endpoint to list all known metrics
+-   Add logging HTTP request and response body sizes
+-   Add process id and reference glob pattern matching
+-   Add cache block list for extensions not to cache
+-   Mod exclude .m3u8 and .mpd files from disk cache by default
+-   Mod replaces x/crypto/acme/autocert with caddyserver/certmagic
+-   Mod exposes ports (Docker desktop)
+-   Fix assigning cleanup rules for diskfs
+-   Fix wrong path for swagger definition
+-   Fix process cleanup on delete, remove empty directories from disk
+-   Fix SRT blocking port on restart (upgrade datarhei/gosrt)
+-   Fix RTMP communication (Blackmagic Web Presenter, thx 235 MEDIA)
+-   Fix RTMP communication (Blackmagic ATEM Mini, datarhei/restreamer#385)
+-   Fix injecting commit, branch, and build info
+-   Fix API metadata endpoints responses
+
 #### Core v16.9.0 > v16.9.1
 
 -   Fix v1 import app